feat(elements): provide vue plugin
