+++ packages/elements-example-vue/src/main.js
@@ -1,9 +1,7 @@
 import { createApp } from 'vue';
-import { addIcons } from '@inovex.de/elements/dist/collection/util/icons';
-import { ICON_PATHS } from '@inovex.de/elements/dist/inovex-elements/ino-icon/index.esm.js';
+import { InoElementsVue } from '@inovex.de/elements-vue';
 import App from './App.vue'
 
-addIcons(ICON_PATHS); // c) (only needed if you want to use our icons)
-
 const app = createApp(App);
+app.use(InoElementsVue);
 app.mount('#app');
